\name{toy_data}

\alias{toy_data}
\alias{toy_genes_gff}
\alias{toy_reads_sam}
\alias{toy_reads_bam}

\title{
  Little helpers for quick access to the toy data included in the
  SplicingGraphs package
}

\description{
  TODO
}

\usage{
toy_genes_gff()
toy_reads_sam()
toy_reads_bam()
}

\author{
  H. Pages
}

\seealso{
  The toy data described here is used in the examples of the following
  man pages:
  \itemize{
    \item \code{\link{splicingGraphs}}: Tools for creating and
          plotting splicing graphs from transcript annotations.

    \item \link{countReads}: Tools for assigning read counts to splicing
          graphs.
  }
}

\examples{
toy_genes_gff()
toy_reads_sam()
toy_reads_bam()
}
